requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
pillow
einops
tqdm
datasets
argparse
omegaconf
matplotlib
transparent-background
opencv-python
timm
rembg
python-dotenv
wandb
accelerate
bitsandbytes
fonts
roboto
font_roboto


# extra index url for xformers with cu118
# 
torch
torchaudio
torchvision
#--extra-index-url https://download.pytorch.org/whl/cu121
# xformers
